CLOUD_FUNCTION/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
google-cloud-bigquery
requests
